+++ index.html
@@ -229,7 +229,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="./js/main.js">
+  <script type="module" src="./js/main.js">
 
   </script>
 </body>
